requirements.txt
django
djangorestframework
djangorestframework_simplejwt
pymysql
mysqlclient
pillow